[TypeScript] fix array enum (#3538)
* better enum support for typescript angular2

* fix enum property for TS

* fix non-array enum datatype for TS

* update build name for TS fetch petstore client

* restore TS fetch enum naming

* [TS fetch]add test cases, fix enum integer naming

+++ samples/client/petstore/typescript-angular/API/Client/StoreApi.ts
@@ -61,9 +61,9 @@ namespace API.Client {
 
             let queryParameters: any = {};
             let headerParams: any = this.extendObj({}, this.defaultHeaders);
-            // verify required parameter 'orderId' is set
-            if (!orderId) {
-                throw new Error('Missing required parameter orderId when calling deleteOrder');
+            // verify required parameter 'orderId' is not null or undefined
+            if (orderId === null || orderId === undefined) {
+                throw new Error('Required parameter orderId was null or undefined when calling deleteOrder.');
             }
             let httpRequestParams: any = {
                 method: 'DELETE',
@@ -113,9 +113,9 @@ namespace API.Client {
 
             let queryParameters: any = {};
             let headerParams: any = this.extendObj({}, this.defaultHeaders);
-            // verify required parameter 'orderId' is set
-            if (!orderId) {
-                throw new Error('Missing required parameter orderId when calling getOrderById');
+            // verify required parameter 'orderId' is not null or undefined
+            if (orderId === null || orderId === undefined) {
+                throw new Error('Required parameter orderId was null or undefined when calling getOrderById.');
             }
             let httpRequestParams: any = {
                 method: 'GET',
